added swagger, logging, testing; minor refactor changes
