--- src/main/java/dev/klimkin/creditconveyor/service/CalculationService.java
@@ -1,12 +1,0 @@
-package dev.klimkin.creditconveyor.service;
-
-import java.math.BigDecimal;
-
-public interface CalculationService {
-    BigDecimal calculateMonthlyRate(BigDecimal rate);
-    BigDecimal calculateAnnuityRatio(Integer term, BigDecimal rate);
-    BigDecimal calculateInsurance(BigDecimal requestedAmount, Integer term);
-    BigDecimal calculateMonthlyPayment(BigDecimal requestedAmount, Integer term, BigDecimal annuityRatio,
-                                       boolean isInsuranceEnabled);
-    BigDecimal calculateTotalAmount(BigDecimal monthlyPayment, Integer term);
-}
