--- src/main/java/dev/klimkin/creditconveyor/service/CreditService.java
@@ -1,9 +1,0 @@
-package dev.klimkin.creditconveyor.service;
-
-import dev.klimkin.creditconveyor.dto.CreditDTO;
-import dev.klimkin.creditconveyor.dto.ScoringDataDTO;
-
-public interface CreditService {
-
-    CreditDTO calculateCredit(ScoringDataDTO scoringDataDTO);
-}
