--- src/main/java/dev/klimkin/creditconveyor/service/OfferService.java
@@ -1,13 +1,0 @@
-package dev.klimkin.creditconveyor.service;
-
-import dev.klimkin.creditconveyor.dto.LoanApplicationRequestDTO;
-import dev.klimkin.creditconveyor.dto.LoanOfferDTO;
-
-
-import java.util.List;
-
-public interface OfferService {
-
-    List<LoanOfferDTO> getLoanOffer(LoanApplicationRequestDTO loanApplicationRequestDTO);
-
-}
