--- src/main/java/dev/klimkin/creditconveyor/util/LoanApplicationRequestErrorException.java
@@ -1,7 +1,0 @@
-package dev.klimkin.creditconveyor.util;
-
-public class LoanApplicationRequestErrorException extends RuntimeException {
-    public LoanApplicationRequestErrorException(String msg){
-        super(msg);
-    }
-}
