--- src/main/java/dev/klimkin/creditconveyor/util/LoanApplicationRequestErrorResponse.java
@@ -1,13 +1,0 @@
-package dev.klimkin.creditconveyor.util;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@AllArgsConstructor
-@Getter
-@Setter
-public class LoanApplicationRequestErrorResponse {
-    private String message;
-    private long timestamp;
-}
