--- src/main/java/dev/klimkin/creditconveyor/util/ScoringDataErrorException.java
@@ -1,7 +1,0 @@
-package dev.klimkin.creditconveyor.util;
-
-public class ScoringDataErrorException extends RuntimeException{
-    public ScoringDataErrorException(String msg){
-        super(msg);
-    }
-}
