--- src/main/java/dev/klimkin/creditconveyor/util/ScoringResultException.java
@@ -1,7 +1,0 @@
-package dev.klimkin.creditconveyor.util;
-
-public class ScoringResultException extends RuntimeException {
-    public ScoringResultException(String message) {
-        super(message);
-    }
-}
