+++ src/main/java/ru/klimkin/creditconveyor/config/SwaggerConfig.java
@@ -1,0 +1,19 @@
+package ru.klimkin.creditconveyor.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .info(
+                        new Info()
+                                .title("credit-conveyor-mvp1")
+                                .version("2.0.0")
+                );
+    }
+}
