+++ src/main/java/ru/klimkin/creditconveyor/service/OfferService.java
@@ -1,0 +1,13 @@
+package ru.klimkin.creditconveyor.service;
+
+import ru.klimkin.creditconveyor.dto.LoanApplicationRequestDTO;
+import ru.klimkin.creditconveyor.dto.LoanOfferDTO;
+
+
+import java.util.List;
+
+public interface OfferService {
+
+    List<LoanOfferDTO> getOfferList(LoanApplicationRequestDTO loanApplicationRequestDTO);
+
+}
