+++ src/main/java/ru/klimkin/creditconveyor/util/LoanApplicationRequestErrorException.java
@@ -1,0 +1,7 @@
+package ru.klimkin.creditconveyor.util;
+
+public class LoanApplicationRequestErrorException extends RuntimeException {
+    public LoanApplicationRequestErrorException(String msg){
+        super(msg);
+    }
+}
