+++ src/main/java/ru/klimkin/creditconveyor/util/LoanApplicationRequestErrorResponse.java
@@ -1,0 +1,13 @@
+package ru.klimkin.creditconveyor.util;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
+public class LoanApplicationRequestErrorResponse {
+    private String message;
+    private long timestamp;
+}
