+++ src/main/java/ru/klimkin/creditconveyor/util/ScoringDataErrorException.java
@@ -1,0 +1,7 @@
+package ru.klimkin.creditconveyor.util;
+
+public class ScoringDataErrorException extends RuntimeException{
+    public ScoringDataErrorException(String msg){
+        super(msg);
+    }
+}
