+++ src/main/java/ru/klimkin/creditconveyor/util/ScoringDataErrorResponse.java
@@ -1,0 +1,13 @@
+package ru.klimkin.creditconveyor.util;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
+public class ScoringDataErrorResponse {
+    private String message;
+    private long timestamp;
+}
