+++ src/main/java/ru/klimkin/creditconveyor/util/ScoringResultException.java
@@ -1,0 +1,7 @@
+package ru.klimkin.creditconveyor.util;
+
+public class ScoringResultException extends RuntimeException {
+    public ScoringResultException(String message) {
+        super(message);
+    }
+}
