--- src/test/java/dev/klimkin/creditconveyor/CreditConveyorApplicationTests.java
@@ -1,13 +1,0 @@
-package dev.klimkin.creditconveyor;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CreditConveyorApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
