+++ src/test/java/ru/klimkin/creditconveyor/CreditConveyorApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.klimkin.creditconveyor;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CreditConveyorApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
